Boss debuff cleanup: -Trojan Squirrel snowballs no longer inflict Chilled -Cursed Coffin: Reduced shadowflame debuff duration -Brain of Cthulhu: Massively decreased Ichor debuff duration, removed most Maso mass debuff applications -Eater of Worlds: Tweaked debuff durations -Small Bees or Beehive projectiles or Subjects no longer inflict Swarming -Skeletron: Hands no longer inflict Dazed -Lethargic attack speed reduction 25% -> 15%
Other boss changes:
-EoC: Reduced servant spawns and max servants slightly
-Coffin: Slightly decreased damage
-All grabs require less mashing to escape, but more mashing has decreased gain
-Eater of Worlds: Reduced sine wave movement amplitude during cursed fire attack (aims at you more)
-Queen Bee: Removed randomness from shot angles

+++ Content/Projectiles/Eternity/Bosses/Deerclops/DeerclopsHand.cs
@@ -32,7 +32,7 @@ namespace FargowiltasSouls.Content.Projectiles.Eternity.Bosses.Deerclops
         }
 
         bool spawned;
-        int oldMash;
+        float oldMash;
 
         public override void AI()
         {
